Add file dropdown feature to the map
fixes: #2437
